📚 docs: add token meta info to token system model (#39101)
* :recycle: refactor token system interface

* :memo: docs: add more token meta info to seeds and map token

* :hammer: chore: update token meta generator scripts

* :rotating_light: chore: fix lint

* :memo: docs: 补充 token 的 meta 信息

* chore: add json in pkg

* chore: update script

* chore: update flag
